nicer app icon detection + usage + new hook
